Create 1738305.yml
Notes: I'm not a fan of doing these one file at a time, I'd rather edit 1 file with several entries.

Also, it would be nice if there was a human-readable way to see what the publication is about. Maybe a `note:` field that isn't rendered, but allows someone looking at the yaml to know what it's about.