Create 20220719-DANCE-CoDaS-Snowmass-Seattle.yml
The DANCE/CoDaS@Snowmass 2022 event is a collaboration between the Dark
Matter and Neutrino Computation Explored (DANCE) collaboration and
the Computational and Data Science for High Energy Physics school.

Advanced software is a critical ingredient to scientific research. Training young researchers in the latest tools and techniques is an essential part of developing the skills required for a successful career both in research and in industry.

DANCE/CoDaS@Snowmass 2022 is aimed at senior graduate students and/or early postdocs (in their first year) who have a solid base of software skills and are interested in developing more advanced skills as well as participating in the Snowmass Community Summer Study. Applicants should be affiliated with a U.S. university (students/postdocs) or laboratory (postdocs).

Specific topics to be covered at the school include:

- The Scientific Python ecosystem for particle physics
- Advanced Data Analysis tools in the Python Ecosystem
- Machine Learning, Bayesian Networks
- Statistical tools for particle physics

The school offers a limited number of young researchers an opportunity to learn these skills from experienced scientists and instructors. Successful applicants will receive travel and lodging support to attend the school.