(jpivarski) ACAT 2024 talk and poster (#2125)
* (jpivarski) ACAT 2024 talk and poster

This is a draft, but it has all of the URLs set up. I also need to check in about the talk because I don't see it on the schedule. (It would be okay by me if it's dropped, but I was never told...)

* the JIT compilation talk is not happening

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: Rob Tuck <robtuck@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>